% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_delete_object_request}
\alias{add_delete_object_request}
\title{Add a delete object request}
\usage{
add_delete_object_request(google_slides_request = NULL, object_id)
}
\arguments{
\item{google_slides_request}{A \code{GoogleSlidesRequest} object, or \code{NULL}.}

\item{object_id}{ID of the object to delete.}
}
\value{
A \code{GoogleSlidesRequest} object with the request appended.
}
\description{
Add a delete object request
}
