% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_delete_table_column_request}
\alias{add_delete_table_column_request}
\title{Add a delete table column request}
\usage{
add_delete_table_column_request(
  google_slides_request = NULL,
  table_object_id,
  row_index,
  column_index
)
}
\arguments{
\item{google_slides_request}{A \code{GoogleSlidesRequest} object, or \code{NULL}.}

\item{table_object_id}{The table to delete a column from.}

\item{row_index}{The 0-based row index.}

\item{column_index}{The 0-based column index.}
}
\value{
A \code{GoogleSlidesRequest} object with the request appended.
}
\description{
Add a delete table column request
}
