% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_insert_table_columns_request}
\alias{add_insert_table_columns_request}
\title{Add an insert table columns request}
\usage{
add_insert_table_columns_request(
  google_slides_request = NULL,
  table_object_id,
  row_index,
  column_index,
  insert_right = TRUE,
  number
)
}
\arguments{
\item{google_slides_request}{A \code{GoogleSlidesRequest} object, or \code{NULL}.}

\item{table_object_id}{The table to insert columns into.}

\item{row_index}{The 0-based row index.}

\item{column_index}{The 0-based column index.}

\item{insert_right}{If \code{TRUE}, insert to the right of the reference cell.}

\item{number}{Number of columns to insert (max 20).}
}
\value{
A \code{GoogleSlidesRequest} object with the request appended.
}
\description{
Add an insert table columns request
}
