% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_insert_table_rows_request}
\alias{add_insert_table_rows_request}
\title{Add an insert table rows request}
\usage{
add_insert_table_rows_request(
  google_slides_request = NULL,
  table_object_id,
  row_index,
  column_index,
  insert_below = TRUE,
  number
)
}
\arguments{
\item{google_slides_request}{A \code{GoogleSlidesRequest} object, or \code{NULL}.}

\item{table_object_id}{The table to insert rows into.}

\item{row_index}{The 0-based row index.}

\item{column_index}{The 0-based column index.}

\item{insert_below}{If \code{TRUE}, insert below the reference cell.}

\item{number}{Number of rows to insert (max 20).}
}
\value{
A \code{GoogleSlidesRequest} object with the request appended.
}
\description{
Add an insert table rows request
}
