% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_replace_all_text_request}
\alias{add_replace_all_text_request}
\title{Add a replace all text request}
\usage{
add_replace_all_text_request(
  google_slides_request = NULL,
  replace_text,
  text,
  match_case = TRUE
)
}
\arguments{
\item{google_slides_request}{A \code{GoogleSlidesRequest} object, or \code{NULL}.}

\item{replace_text}{Replacement text.}

\item{text}{Text to search for.}

\item{match_case}{Whether to match case.}
}
\value{
A \code{GoogleSlidesRequest} object with the request appended.
}
\description{
Add a replace all text request
}
