% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_create_shape}
\alias{build_create_shape}
\title{Building create shape request}
\usage{
build_create_shape(shapeType = NULL, pageElementPropert = NULL,
  objectId = NULL)
}
\arguments{
\item{shapeType}{A list that contains the type of shape that is to be drawn}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}

\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the page_element_property function in this package.}
}
\description{
This function builds up the request for creating shapes within Googleslides via the API.
}

