% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_create_slide}
\alias{build_create_slide}
\title{Building create slide request}
\usage{
build_create_slide(no_of_slides = 1, insertionIndex = NULL,
  layoutId = NULL, predefinedLayout = NULL, objectId = NULL,
  requests_list = NULL)
}
\description{
Building create slide request
}

