% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_create_slide_page}
\alias{build_create_slide_page}
\title{Building create slide request}
\usage{
build_create_slide_page(no_of_slides = 1, insertionIndex = NULL,
  layoutId = NULL, predefinedLayout = NULL, objectId = NULL,
  requests_list = NULL)
}
\arguments{
\item{no_of_slides}{A number to indicate the number of slides that is to be added}

\item{insertionIndex}{A numeric vector on where the slide is to be added}

\item{layoutId}{A character vector that provides guidance on which layout the new slide is to follow}

\item{predefinedLayout}{A character vector that provides guidance on which layout the new slide
is to follow. The ones declared here}

\item{objectId}{A character vector that is to be used to give names to new slides created via the
slides API}

\item{requests_list}{A list of requests that is to be passed to the post_batchUpdate function.}
}
\description{
Building create slide request
}

