% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_delete_object}
\alias{build_delete_object}
\title{Building delete object request}
\usage{
build_delete_object(objectId = NULL, requests_list = NULL)
}
\arguments{
\item{objectId}{A character vector of object ids that is to be deleted from the slides}

\item{requests_list}{A list of requests that is to be passed to the post_batchUpdate function.}
}
\description{
Building delete object request
}

