% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_delete_text}
\alias{build_delete_text}
\title{Building delete text request}
\usage{
build_delete_text(objectId = NULL, rowIndex = NULL, columnIndex = NULL,
  startIndex = NULL, endIndex = NULL, type = "ALL",
  requests_list = NULL)
}
\arguments{
\item{objectId}{A character vector of objects to insert text into. You can only insert text in
tables and shapes}

\item{rowIndex}{A numeric vector of rows to insert the text into}

\item{columnIndex}{A numeric vector of columns to insert the text into}

\item{startIndex}{The optional zero-based index of the beginning of the collection.
Required for SPECIFIC_RANGE and FROM_START_INDEX ranges.}

\item{endIndex}{The optional zero-based index of the end of the collection.
Required for SPECIFIC_RANGE delete mode.}

\item{type}{The type of range. Can be the following RANGE_TYPE_UNSPECIFIED, FIXED_RANGE,
FROM_START_INDEX, ALL}

\item{requests_list}{A list of requests that is to be passed to the post_batchUpdate function.}
}
\description{
Building delete text request
}

