% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_insert_text}
\alias{build_insert_text}
\title{Building insert text request}
\usage{
build_insert_text(objectId = NULL, rowIndex = NULL, columnIndex = NULL,
  text = NULL, insertionIndex = NULL, requests_list = NULL)
}
\arguments{
\item{objectId}{A character vector of objects to insert text into. You can only insert text in
tables and shapes}

\item{rowIndex}{A numeric vector of rows to insert the text into}

\item{columnIndex}{A numeric vector of columns to insert the text into}

\item{text}{A character vector of text which is to be inserted into the shape or table}

\item{insertionIndex}{A numeric vector which indicate the starting point of how the text
is to be inserted}

\item{requests_list}{A list of requests that is to be passed to the post_batchUpdate function.}
}
\description{
Building insert text request
}

