% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_update_slides_position}
\alias{build_update_slides_position}
\title{Building update slides position request}
\usage{
build_update_slides_position(slideObjectIds = NULL, insertionIndex = NULL,
  requests_list = NULL)
}
\arguments{
\item{slideObjectIds}{List of Character Vector.}

\item{insertionIndex}{Numeric Vector. This is where the slides selected in
slideobjectids parameter is inserted into}
}
\description{
Building update slides position request
}

