% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchUpdate.R
\name{commit_to_slides}
\alias{commit_to_slides}
\title{Commit changes to slides}
\usage{
commit_to_slides(id, google_slide_request)
}
\arguments{
\item{id}{The presentation ID.}

\item{google_slide_request}{A \code{GoogleSlidesRequest} object created by the
\verb{add_*_request()} functions.}
}
\value{
A list containing the API response with any reply data.
}
\description{
Send a batch of accumulated requests to the Google Slides API.
}
\examples{
\dontshow{if (gs_has_token()) withAutoprint(\{ # examplesIf}
request <- add_create_slide_page_request()
commit_to_slides("<presentation-id>", request)
\dontshow{\}) # examplesIf}
}
