% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_image}
\alias{create_image}
\title{Insert image to a page in a slide}
\usage{
create_image(id = NULL, url = NULL, pageElementProperty = NULL,
  objectId = NULL)
}
\arguments{
\item{id}{ID of the presenation slide}
}
\description{
Insert image to a page in a slide
}

