% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_shape}
\alias{create_shape}
\title{Insert shape to a page in a slide}
\usage{
create_shape(id = NULL, shapeType = NULL, pageElementProperty = NULL,
  objectId = NULL)
}
\arguments{
\item{id}{ID of the presenation slide}

\item{shapeType}{A character vector that contains the shape type for the new shape that is to be created}

\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the page_element_property function in this package. IT IS COMPULSORY TO ADD WIDTH AND HEIGHT AS
PART OF THE page_element_property}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}
}
\description{
Insert shape to a page in a slide
}

