% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_slide_page}
\alias{create_slide_page}
\title{Add page in slides}
\usage{
create_slide_page(id = NULL, no_of_slides = 1, insertionIndex = NULL,
  layoutId = NULL, predefinedLayout = NULL, objectId = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Add page in slides
}

