% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataframe_convert}
\alias{dataframe_convert}
\title{Convert dataframe to dataframe with rows and columns}
\usage{
dataframe_convert(data = NULL, headers = TRUE)
}
\arguments{
\item{data}{Dataframe of the dataset that is to be converted so that it can be used with the google slides API}

\item{headers}{Boolean to indicate whether to convert taking in mind of the headers or not}
}
\description{
Convert dataframe to dataframe with rows and columns
}

