% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{delete_object}
\alias{delete_object}
\title{Delete object in slides}
\usage{
delete_object(id = NULL, objectId = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Delete object in slides
}

