% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Generate endpoint for the Google Slides API
Temporarily available}
\usage{
get_endpoint(typeOfEndpoint = "slides.endpoint.get", id = NULL)
}
\description{
Generate endpoint for the Google Slides API
Temporarily available
}

