% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_google_drive_urls}
\alias{get_google_drive_urls}
\title{Get the list of google drive url}
\usage{
get_google_drive_urls(imageId)
}
\arguments{
\item{imageId}{ID of the image on Google Drive}
}
\description{
Get the list of google drive url
}

