% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{get_slides_properties}
\alias{get_slides_properties}
\title{Get Google Slides properties}
\usage{
get_slides_properties(id)
}
\arguments{
\item{id}{Presentation ID.}
}
\value{
A list containing the presentation properties.
}
\description{
Retrieve the full properties of a Google Slides presentation.
}
\examples{
\dontshow{if (gs_has_token()) withAutoprint(\{ # examplesIf}
props <- get_slides_properties("<presentation-id>")
\dontshow{\}) # examplesIf}
}
