% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_auth}
\alias{gs_auth}
\title{Authorize rgoogleslides}
\usage{
gs_auth(
  email = gargle::gargle_oauth_email(),
  path = NULL,
  scopes = c("https://www.googleapis.com/auth/presentations",
    "https://www.googleapis.com/auth/drive.readonly"),
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default(),
  token = NULL
)
}
\arguments{
\item{email}{Optional. If specified, \code{email} can take several different
forms:
\itemize{
\item \code{"jane@gmail.com"}, an actual email address. If found in the cache,
the associated token is loaded. Otherwise, the OAuth flow is
initiated using this email as a hint.
\item \code{"*@example.com"}, a domain-only pattern. Any cached token matching
this pattern is loaded.
\item \code{TRUE} means that a cached token is loaded if exactly one match is
found.
\item \code{FALSE} or \code{NA} mean to ignore the cache and always initiate a new
OAuth flow.
}}

\item{path}{Path to a service account JSON file.}

\item{scopes}{A character vector of scopes for the access request.}

\item{cache}{Specifies the OAuth token cache. Set \code{FALSE} to disable.}

\item{use_oob}{Whether to use out-of-band authentication.}

\item{token}{A token to use directly, an object of class \link[httr:Token-class]{httr::Token2.0},
or a pre-configured request as returned by \code{gs_token()}.}
}
\value{
Called for its side effect of storing a token for later use.
Returns the token invisibly.
}
\description{
Authorize rgoogleslides to access the Google Slides API on your behalf.
This function is a wrapper around \code{\link[gargle:token_fetch]{gargle::token_fetch()}}.
}
\details{
Most users, most of the time, do not need to call \code{gs_auth()} explicitly --
it is triggered automatically by the first action that requires
authorization. Even when called, the default arguments often suffice.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# trigger interactive OAuth flow
gs_auth()

# use a specific email
gs_auth(email = "jane@example.com")

# use a service account
gs_auth(path = "path/to/service-account.json")
\dontshow{\}) # examplesIf}
}
\seealso{
Other auth: 
\code{\link{gs_auth_configure}()},
\code{\link{gs_deauth}()},
\code{\link{gs_has_token}()},
\code{\link{gs_oauth_client}()},
\code{\link{gs_token}()}
}
\concept{auth}
