% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_auth_configure}
\alias{gs_auth_configure}
\title{Edit auth configuration}
\usage{
gs_auth_configure(client, path)
}
\arguments{
\item{client}{A Google OAuth client, as produced by
\code{\link[gargle:gargle_oauth_client_from_json]{gargle::gargle_oauth_client()}}.}

\item{path}{Path to a JSON file with client credentials.}
}
\value{
None, called for side effects.
}
\description{
Edit auth configuration
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
gs_auth_configure(path = "/path/to/oauth-client.json")
\dontshow{\}) # examplesIf}
}
\seealso{
Other auth: 
\code{\link{gs_auth}()},
\code{\link{gs_deauth}()},
\code{\link{gs_has_token}()},
\code{\link{gs_oauth_client}()},
\code{\link{gs_token}()}
}
\concept{auth}
