% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_deauth}
\alias{gs_deauth}
\title{Clear Google Slides credentials}
\usage{
gs_deauth()
}
\value{
None, called for side effects.
}
\description{
Put rgoogleslides into a de-authorized state. After calling \code{gs_deauth()},
any request that requires authorization will fail.
}
\examples{
gs_deauth()
}
\seealso{
Other auth: 
\code{\link{gs_auth}()},
\code{\link{gs_auth_configure}()},
\code{\link{gs_has_token}()},
\code{\link{gs_oauth_client}()},
\code{\link{gs_token}()}
}
\concept{auth}
