% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_has_token}
\alias{gs_has_token}
\title{Is there a token on hand?}
\usage{
gs_has_token()
}
\value{
\code{TRUE} if a token is available, \code{FALSE} otherwise.
}
\description{
Reports whether rgoogleslides has stored a token, ready for use in
downstream requests.
}
\examples{
gs_has_token()
}
\seealso{
Other auth: 
\code{\link{gs_auth}()},
\code{\link{gs_auth_configure}()},
\code{\link{gs_deauth}()},
\code{\link{gs_oauth_client}()},
\code{\link{gs_token}()}
}
\concept{auth}
