% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_oauth_client}
\alias{gs_oauth_client}
\title{Retrieve the configured OAuth client}
\usage{
gs_oauth_client()
}
\value{
A gargle OAuth client object, or \code{NULL} if not configured.
}
\description{
Retrieve the configured OAuth client
}
\examples{
gs_oauth_client()
}
\seealso{
Other auth: 
\code{\link{gs_auth}()},
\code{\link{gs_auth_configure}()},
\code{\link{gs_deauth}()},
\code{\link{gs_has_token}()},
\code{\link{gs_token}()}
}
\concept{auth}
