% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_auth.R
\name{gs_token}
\alias{gs_token}
\title{Produce a configured token}
\usage{
gs_token()
}
\value{
An \code{\link[httr:config]{httr::config()}} object with the current credential, or \code{NULL}
if auth is not active.
}
\description{
For internal use or for other packages that wish to use the same
credentials as rgoogleslides. Returns a token pre-processed into a
suitable form for use with \code{\link[httr:config]{httr::config()}}.
}
\examples{
\dontshow{if (gs_has_token()) withAutoprint(\{ # examplesIf}
gs_token()
\dontshow{\}) # examplesIf}
}
\seealso{
Other auth: 
\code{\link{gs_auth}()},
\code{\link{gs_auth_configure}()},
\code{\link{gs_deauth}()},
\code{\link{gs_has_token}()},
\code{\link{gs_oauth_client}()}
}
\concept{auth}
