% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{insert_text}
\alias{insert_text}
\title{Insert text to shapes/tables}
\usage{
insert_text(id = NULL, objectId = NULL, rowIndex = NULL,
  columnIndex = NULL, text = NULL, insertionIndex = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Insert text to shapes/tables
}

