% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.google_slide_request}
\alias{is.google_slide_request}
\title{Check if the object is a GoogleSlidesRequest}
\usage{
is.google_slide_request(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} is a \code{GoogleSlidesRequest}, \code{FALSE} otherwise.
}
\description{
Check if the object is a GoogleSlidesRequest
}
