% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.page_element_property}
\alias{is.page_element_property}
\title{Check if the object is a PageElementProperty}
\usage{
is.page_element_property(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} is a \code{PageElementProperty}, \code{FALSE} otherwise.
}
\description{
Check if the object is a PageElementProperty
}
