% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonProperties.R
\name{page_element_property}
\alias{page_element_property}
\title{Common Property: Page Element Property}
\usage{
page_element_property(
  page_object_id = NULL,
  width_magnitude = NULL,
  height_magnitude = NULL,
  scale_x = NULL,
  scale_y = NULL,
  shear_x = NULL,
  shear_y = NULL,
  translate_x = NULL,
  translate_y = NULL,
  width_unit = "PT",
  height_unit = "PT",
  transform_unit = "PT"
)
}
\arguments{
\item{page_object_id}{The slide page ID.}

\item{width_magnitude}{Width of the element.}

\item{height_magnitude}{Height of the element.}

\item{scale_x}{Horizontal scale factor.}

\item{scale_y}{Vertical scale factor.}

\item{shear_x}{Horizontal shear factor.}

\item{shear_y}{Vertical shear factor.}

\item{translate_x}{Horizontal translation.}

\item{translate_y}{Vertical translation.}

\item{width_unit}{Unit for width. Default \code{"PT"}.}

\item{height_unit}{Unit for height. Default \code{"PT"}.}

\item{transform_unit}{Unit for transforms. Default \code{"PT"}.}
}
\value{
A \code{PageElementProperty} R6 object.
}
\description{
Build a page element property for positioning and sizing elements on a
slide.
}
\examples{
prop <- page_element_property("slide-id-1", 200, 300)
}
