% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchUpdate.R
\name{post_batchUpdate}
\alias{post_batchUpdate}
\title{Send a POST request to do a batch update on the slides}
\usage{
post_batchUpdate(id = NULL, requests_list = NULL)
}
\description{
Send a POST request to do a batch update on the slides
}

