% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{replace_all_text}
\alias{replace_all_text}
\title{Replace all text within the slide}
\usage{
replace_all_text(id = NULL, replaceText = NULL, text = NULL,
  matchCase = TRUE)
}
\arguments{
\item{id}{ID of the presentation slide}

\item{replaceText}{A character vector of text that would replace the ones in the text parameter.
The order of the replaceText matters}

\item{text}{A character vector of text that would replaced by the ones in the replaceText parameter.
The order of the text matters}

\item{matchCase}{A boolean that takes in only TRUE or FALSE only. This would be applied across all
requests}
}
\description{
Replace all text within the slide
}

