% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgoogleslides.R
\docType{package}
\name{rgoogleslides-package}
\alias{rgoogleslides}
\alias{rgoogleslides-package}
\title{rgoogleslides: R Interface to the 'Google Slides' API}
\description{
Provides an R interface to the 'Google Slides' API \url{https://developers.google.com/slides/api}. Create presentations, add slides, insert text, tables, images, and videos programmatically from R. Designed for reproducible reporting workflows where analysis results need to be pushed directly into 'Google Slides' presentations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eivicent/rgoogleslides}
  \item Report bugs at \url{https://github.com/eivicent/rgoogleslides/issues}
}

}
\author{
\strong{Maintainer}: Vicent Boned \email{vicent.boned@gmail.com} [contributor, copyright holder]

Authors:
\itemize{
  \item Hairizuan Bin Noorazman \email{hairizuanbinnoorazman@gmail.com}
}

}
\keyword{internal}
