% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{update_slides_position}
\alias{update_slides_position}
\title{Update slides position}
\usage{
update_slides_position(id = NULL, slideObjectIds = NULL,
  insertionIndex = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Update slides position
}

