% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_create_image_request}
\alias{add_create_image_request}
\title{Add a create image request}
\usage{
add_create_image_request(google_slides_request = NULL,
  page_element_property, url, object_id = NULL)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{page_element_property}{A list that contains a page element property. The page element is to be
generated by the page_element_property function in this package.}

\item{url}{A character vector container an image url that is to be used to add image to the slides}

\item{object_id}{(Optional) A character vector to name the object created instead of leaving it to Google}
}
\description{
Add a create image request
}
