% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_create_slide_page_request}
\alias{add_create_slide_page_request}
\title{Add a create slide request}
\usage{
add_create_slide_page_request(google_slides_request = NULL,
  insertion_index = NULL, layout_id = NULL, predefined_layout = NULL,
  object_id = NULL)
}
\arguments{
\item{google_slides_request}{A Google Slides Request object which is used to manage requests to the API}

\item{insertion_index}{A numeric vector on where the slide is to be added}

\item{layout_id}{A character vector that provides guidance on which layout the new slide is to follow}

\item{predefined_layout}{A character vector that provides guidance on which layout the new slide
is to follow. The ones declared here}

\item{object_id}{A character vector that is to be used to give names to new slides created via the
slides API}
}
\description{
Add a create slide request
}

