% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_create_table_request}
\alias{add_create_table_request}
\title{Add a create table request}
\usage{
add_create_table_request(google_slides_request = NULL, page_element_property,
  rows, columns, object_id = NULL)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{page_element_property}{A list that contains a page element property. The page element is to be
generated by the page_element_property function in this package.}

\item{rows}{A numeric vector with the row index of the data to be filled in the table. Only for tables}

\item{columns}{A numeric vector with the column index of the data to be filled in the table. Only for tables}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}
}
\description{
Add a create table request
}

