% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_create_video_request}
\alias{add_create_video_request}
\title{Add a create video request}
\usage{
add_create_video_request(google_slides_request = NULL, id,
  page_element_property, object_id = NULL)
}
\arguments{
\item{google_slides_request}{A Google Slides Request object which is used to manage requests to the API}

\item{id}{The video source's unique identifier for this video. e.g. For YouTube
video https://www.youtube.com/watch?v=7U3axjORYZ0 , the ID is 7U3axjORYZ0.}

\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the page_element_property function in this package.}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}
}
\description{
Add a create video request
}

