% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_delete_object_request}
\alias{add_delete_object_request}
\title{Add a delete object request}
\usage{
add_delete_object_request(google_slides_request = NULL, object_id = NULL)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{object_id}{A character vector of object ids that is to be deleted from the slides}
}
\description{
Add a delete object request
}
