% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_delete_table_row_request}
\alias{add_delete_table_row_request}
\title{Add a delete table row request}
\usage{
add_delete_table_row_request(google_slides_request = NULL,
  table_object_id, row_index, column_index)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{table_object_id}{The table to delete a row from.}

\item{row_index}{The 0-based row index.}

\item{column_index}{The 0-based column index.}
}
\description{
Deletes a row from a table.
}
