% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_delete_text_request}
\alias{add_delete_text_request}
\title{Add a delete text request}
\usage{
add_delete_text_request(google_slides_request = NULL, object_id = NULL,
  row_index = NULL, column_index = NULL, start_index = NULL,
  end_index = NULL, type = "ALL")
}
\arguments{
\item{google_slides_request}{A Google Slides Request object which is used to manage requests to the API}

\item{object_id}{A character vector of objects to insert text into. You can only insert text in
tables and shapes}

\item{row_index}{A numeric vector of row to insert the text into}

\item{column_index}{A numeric vector of column to insert the text into}

\item{start_index}{The optional zero-based index of the beginning of the collection.
Required for SPECIFIC_RANGE and FROM_START_INDEX ranges.}

\item{end_index}{The optional zero-based index of the end of the collection.
Required for SPECIFIC_RANGE delete mode.}

\item{type}{The type of range. Can be the following RANGE_TYPE_UNSPECIFIED, FIXED_RANGE,
FROM_START_INDEX, ALL}
}
\description{
Add a delete text request
}

