% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_insert_table_columns_request}
\alias{add_insert_table_columns_request}
\title{Add an insert table column request}
\usage{
add_insert_table_columns_request(google_slides_request = NULL,
  table_object_id, row_index, column_index, insert_right = TRUE, number)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{table_object_id}{The table to insert rows into.}

\item{row_index}{The 0-based row index.}

\item{column_index}{The 0-based column index.}

\item{insert_right}{Whether to insert new columns to the right of the reference cell location. If True, cells will
be inserted to the right of the cell reference. If False, cells will be inserted to the left of the cell reference.}

\item{number}{The number of rows to be inserted. Maximum 20 per request.}
}
\description{
Add an insert table column request
}
