% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_insert_text_request}
\alias{add_insert_text_request}
\title{Add an insert text request}
\usage{
add_insert_text_request(google_slides_request = NULL, object_id,
  row_index = NULL, column_index = NULL, text, insertion_index = NULL)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{object_id}{The ID of the object where the text is to be inserted into}

\item{row_index}{(Optional) A numeric vector of row to insert the text into. This value is only
optional for shapes}

\item{column_index}{(Optional) A numeric vector of column to insert the text into. This value is only
optional for shapes}

\item{text}{A character vector of text which is to be inserted into the shape or table}

\item{insertion_index}{(Optional) A numeric vector which indicate the starting point of how the text
is to be inserted}
}
\description{
Add an insert text request
}

