% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_replace_all_text_request}
\alias{add_replace_all_text_request}
\title{Add a replace all text request}
\usage{
add_replace_all_text_request(google_slides_request = NULL,
  replace_text = NULL, text = NULL, match_case = TRUE)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{replace_text}{A character vector of text that would replace the ones in the text parameter.
The order of the replaceText matters}

\item{text}{A character vector of text that would replaced by the ones in the replaceText parameter.
The order of the text matters}

\item{match_case}{(Optional) A boolean that takes in only TRUE or FALSE only. This would be applied across all
requests}
}
\description{
Add a replace all text request
}

