% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{add_update_slides_position_request}
\alias{add_update_slides_position_request}
\title{Add an update slides position request}
\usage{
add_update_slides_position_request(
  google_slides_request = NULL,
  slide_object_ids = NULL,
  insertion_index = NULL
)
}
\arguments{
\item{google_slides_request}{(Optional) A Google Slides Request object which is used to manage requests to the API}

\item{slide_object_ids}{A character vector of slide ids}

\item{insertion_index}{Numeric Vector. This is where the slides selected in
slideobjectids parameter is inserted into. The slides would be inserted based on before the arrangement
before the move.}
}
\description{
Add an update slides position request
}
