% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authorize}
\alias{authorize}
\title{Authorize R package to access Google Slides API}
\usage{
authorize(client_id = getOption("slides.client.id"),
  client_secret = getOption("slides.client.secret"))
}
\arguments{
\item{client_id}{OAuth client ID. This is obtained from Google API Credentials}

\item{client_secret}{OAuth client secret. This is obtained from Google API Credentials}
}
\description{
This is a function to authorize the R package to access the Googleslides API. If no
client.id and client.secret is provided, the package would provide predefined values.
}
