% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_create_image}
\alias{build_create_image}
\title{Building create image request}
\usage{
build_create_image(url = NULL, pageElementProperty = NULL,
  objectId = NULL, requests_list = NULL)
}
\arguments{
\item{url}{A character vector container a list of image urls that is to be used to add image to
the slides}

\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the common_page_element_property function in this package.}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}
}
\description{
Building create image request
}

