% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_create_table}
\alias{build_create_table}
\title{Building create table request}
\usage{
build_create_table(pageObjectId = NULL, rows = NULL, columns = NULL,
  objectId = NULL, requests_list = NULL)
}
\description{
Building create table request
}

