% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_create_table}
\alias{build_create_table}
\title{Building create table request}
\usage{
build_create_table(pageElementProperty = NULL, rows = NULL,
  columns = NULL, objectId = NULL, requests_list = NULL)
}
\arguments{
\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the common_page_element_property function in this package.}

\item{rows}{A numeric vector with the row index of the data to be filled in the table. Only for tables}

\item{columns}{A numeric vector with the column index of the data to be filled in the table. Only for tables}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}

\item{requests_list}{A list of requests that is to be passed to the post_batchUpdate function.}
}
\description{
Building create table request
}

