% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_replace_all_text}
\alias{build_replace_all_text}
\title{Building replace all text request}
\usage{
build_replace_all_text(requests_list = NULL, replaceText = NULL,
  text = NULL, matchCase = TRUE)
}
\description{
Building replace all text request
}

