% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_validity}
\alias{check_validity}
\title{Convenience function to return a value if the value is NA}
\usage{
check_validity(value)
}
\arguments{
\item{value}{Value to check if the value is valid. If value is NA, return as NULL instead.}
}
\description{
A function that checks and ensure that the value only returns null or a number.
This function can only check one value at a time.
}

