% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchUpdate.R
\name{commit_to_slides}
\alias{commit_to_slides}
\title{Commit changes to slides}
\usage{
commit_to_slides(id, google_slide_request)
}
\arguments{
\item{id}{The slide id of the googleslides to be updated}

\item{google_slide_request}{The google slide request to be applied to the slides}
}
\description{
Commit changes to slides
}

