% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native.R
\name{create_data_table}
\alias{create_data_table}
\title{Create and upload data into a table in slides}
\usage{
create_data_table(id = NULL, pageObjectId = NULL, data = NULL,
  headers = TRUE)
}
\arguments{
\item{pageObjectId}{A character vector with the page on which the table will be created}

\item{data}{Dataframe of the dataset that is to be converted so that it can be used with the google slides API}

\item{headers}{Boolean to indicate whether to convert taking in mind of the headers or not}
}
\description{
Create and upload data into a table in slides
}

