% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_image}
\alias{create_image}
\title{Insert image to a page in a slide}
\usage{
create_image(id = NULL, url = NULL, pageElementProperty = NULL,
  objectId = NULL)
}
\arguments{
\item{id}{ID of the presenation slide}

\item{url}{A character vector container a list of image urls that is to be used to add image to
the slides}

\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the page_element_property function in this package.}

\item{objectId}{(Optional) A character vector to name the object created instead of leaving it to Google}
}
\description{
Insert image to a page in a slide
}

