% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_shape}
\alias{create_shape}
\title{Insert shape to a page in a slide}
\usage{
create_shape(id = NULL, shapeType = NULL, pageElementProperty = NULL,
  objectId = NULL)
}
\arguments{
\item{id}{ID of the presenation slide}
}
\description{
Insert shape to a page in a slide
}

