% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_slide}
\alias{create_slide}
\title{Add page in slides}
\usage{
create_slide(id = NULL, no_of_slides = 1, insertionIndex = NULL,
  layoutId = NULL, predefinedLayout = NULL, objectId = NULL)
}
\description{
Add page in slides
}

