% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_slide}
\alias{create_slide}
\title{Add page in slides}
\usage{
create_slide(id = NULL, no_of_slides = 1, insertionIndex = NULL,
  layoutId = NULL, predefinedLayout = NULL, objectId = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}

\item{no_of_slides}{A number to indicate the number of slides that is to be added}

\item{insertionIndex}{A numeric vector on where the slide is to be added}

\item{layoutId}{A character vector that provides guidance on which layout the new slide is to follow}

\item{predefinedLayout}{A character vector that provides guidance on which layout the new slide
is to follow. The ones declared here}

\item{objectId}{A character vector that is to be used to give names to new slides created via the
slides API}
}
\description{
Add page in slides
}

