% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_slides}
\alias{create_slides}
\title{Create a new googleslide}
\usage{
create_slides(title = NULL, full_response = FALSE)
}
\arguments{
\item{title}{Title of the presentation slide}

\item{full_response}{Parameter to decide whether to return the full response or just the presentation ID}
}
\description{
Create a new googleslide
}

