% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_table}
\alias{create_table}
\title{Create table to a page in a slide}
\usage{
create_table(id = NULL, pageObjectId = NULL, rows = NULL,
  columns = NULL, objectId = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Create table to a page in a slide
}

