% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{create_table}
\alias{create_table}
\title{Create table to a page in a slide}
\usage{
create_table(id = NULL, pageElementProperty = NULL, rows = NULL,
  columns = NULL, objectId = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}

\item{pageElementProperty}{A list that contains a page element property. The page element is to be
generated by the common_page_element_property function in this package.}

\item{rows}{A numeric vector with the row index of the data to be filled in the table. Only for tables}

\item{columns}{A numeric vector with the column index of the data to be filled in the table. Only for tables}

\item{objectId}{Optional Field. A character vector to name the object created instead of leaving it to Google}
}
\description{
Create table to a page in a slide
}

