% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{delete_text}
\alias{delete_text}
\title{Delete text to shapes/tables}
\usage{
delete_text(id = NULL, objectId = NULL, rowIndex = NULL,
  columnIndex = NULL, startIndex = NULL, endIndex = NULL, type = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Delete text to shapes/tables
}

