% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{get_slide_page_properties}
\alias{get_slide_page_properties}
\title{Get a single page of a Google Slides property}
\usage{
get_slide_page_properties(id = NULL, pageObjectId = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}

\item{pageObjectId}{The page ID of the presentation slide}
}
\description{
Get a single page of a Google Slides property
}

