% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{get_slides_properties}
\alias{get_slides_properties}
\title{Get Google Slides Properties}
\usage{
get_slides_properties(id, as = NULL)
}
\arguments{
\item{id}{ID of the presentation slide}

\item{as}{desired type of output: raw, text or parsed. Passed to
\code{\link[httr]{content}}}
}
\description{
Get Google Slides Properties
}
