% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.google_slide_request}
\alias{is.google_slide_request}
\title{Check if the object is a google slide request object}
\usage{
is.google_slide_request(x)
}
\description{
Check if the object is a google slide request object
}

