% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.page_element_property}
\alias{is.page_element_property}
\title{Check if the object is a google slide request object}
\usage{
is.page_element_property(x)
}
\arguments{
\item{x}{A page_element_property object created from the rgoogleslides package}
}
\description{
Check if the object is a google slide request object
}
