% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonProperties.R
\name{page_element_property}
\alias{page_element_property}
\title{Common Property: Page Element Property}
\usage{
page_element_property(pageObjectId = NULL, width_magnitude = NULL,
  height_magnitude = NULL, scaleX = NULL, scaleY = NULL, shearX = NULL,
  shearY = NULL, translateX = NULL, translateY = NULL,
  width_unit = "EMU", height_unit = "EMU", transform_unit = "EMU")
}
\arguments{
\item{pageObjectId}{A list of character vectors that contains the page id of the slides}

\item{width_magnitude}{(Optional) A list of numeric vector that contains the width of the component}

\item{height_magnitude}{(Optional) A list of numeric vector that contains the height of the component}

\item{scaleX}{(Optional) A list of numeric vector that tells how the object is to be sized}

\item{scaleY}{(Optional) A list of numeric vector that tells how the object is to be sized}

\item{shearX}{(Optional) A list of numeric vector that tells how the object is to be sheared}

\item{shearY}{(Optional) A list of numeric vector that tells how the object is to be sheared}

\item{transformX}{(Optional) A list of numeric vector that tells how the object is to be located}

\item{transformY}{(Optional) A list of numeric vector that tells how the object is to be located}
}
\description{
This property is repeated in various of the request. Hence, to reduce the number of
times this would appear in the code base, a function is created to handle the creation of this
list.
}

