% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{replace_all_text}
\alias{replace_all_text}
\title{Replace all text within the slide}
\usage{
replace_all_text(id = NULL, replaceText = NULL, text = NULL,
  matchCase = TRUE)
}
\arguments{
\item{id}{ID of the presentation slide}
}
\description{
Replace all text within the slide
}

